refactor: split rates desk sections
