+++ web/components/rates/RatesDesk.tsx
@@ -1,61 +1,23 @@
-import type { components } from "@/lib/types";
 import type { ReactNode } from "react";
 import { RatesCurveChart } from "./RatesCurveChart";
 import styles from "./RatesDesk.module.css";
 import { RatesScorecard } from "./RatesScorecard";
 import { RatesSection } from "./RatesSection";
 import { fmtSigned, fmtValue, statusLabel, toFiniteNumber } from "./format";
-
-type Snapshot = components["schemas"]["RatesSnapshotResponse"];
-type SummaryTile = components["schemas"]["RatesSummaryTile"];
-type SlopeMetric = components["schemas"]["RatesSlopeMetric"];
-type Scorecard = components["schemas"]["RatesScorecard"];
-type Decomposition = NonNullable<Snapshot["decomposition"]>;
-type DecompositionAttribution =
-  components["schemas"]["RatesDecompositionAttribution"];
-type Policy = NonNullable<Snapshot["policy"]>;
-type PolicyPathPoint = components["schemas"]["RatesPolicyPathPoint"];
-type Supply = NonNullable<Snapshot["supply"]> & {
-  recent_auctions?: SupplyAuction[];
-  fiscal?: SummaryTile[];
-  supply_read?: string | null;
-};
-type SupplyAuction = {
-  cusip: string;
-  security_type: string;
-  security_term: string;
-  auction_date: string;
-  issue_date?: string | null;
-  offering_amount?: number | null;
-  high_rate?: number | null;
-  bid_to_cover?: number | null;
-  direct_bidder_pct?: number | null;
-  indirect_bidder_pct?: number | null;
-  primary_dealer_pct?: number | null;
-  tail_indicator?: string | null;
-  source_url?: string | null;
-  status?: string;
-};
-type Positioning = NonNullable<Snapshot["positioning"]> & {
-  details?: PositioningDetail[];
-  positioning_read?: string | null;
-};
-type PositioningDetail = {
-  contract_code: string;
-  contract_name: string;
-  tenor_bucket: string;
-  obs_date?: string | null;
-  release_date?: string | null;
-  open_interest?: number | null;
-  dealer_net?: number | null;
-  dealer_net_pct_oi?: number | null;
-  asset_mgr_net?: number | null;
-  asset_mgr_net_pct_oi?: number | null;
-  lev_money_net?: number | null;
-  lev_money_net_pct_oi?: number | null;
-  source_url?: string | null;
-  status?: string;
-};
+import { PolicySection } from "./sections/PolicySection";
+import { PositioningSection } from "./sections/PositioningSection";
+import { SupplySection } from "./sections/SupplySection";
+import type {
+  Decomposition,
+  DecompositionAttribution,
+  Policy,
+  Positioning,
+  Scorecard,
+  SlopeMetric,
+  Snapshot,
+  SummaryTile,
+  Supply,
+} from "./types";
 
 const NAV = [
   ["summary", "Summary"],
@@ -166,292 +128,6 @@ function Tile({ tile }: { tile: SummaryTile }) {
         {fmtSigned(tile.delta_1d, deltaUnit(tile))}
       </small>
     </article>
-  );
-}
-
-function policyToneClass(stance: string | undefined): string {
-  if (stance === "HIKE") return styles.deltaPositive;
-  if (stance === "CUT") return styles.deltaNegative;
-  return styles.deltaNeutral;
-}
-
-function latestPolicyRows(policy: Policy) {
-  return [
-    ["Target range", policy.target_range ?? "n/a"],
-    ["EFFR", fmtValue(policy.effr, "%", 2)],
-    ["SOFR", fmtValue(policy.sofr, "%", 2)],
-    [
-      "Last meeting",
-      policy.last_meeting?.label
-        ? `${policy.last_meeting.label} · ${policy.last_meeting.action ?? "n/a"}`
-        : "n/a",
-    ],
-    ["Vote split", policy.last_meeting?.vote_split ?? "n/a"],
-  ];
-}
-
-function fmtPolicyMetric(value: unknown, unit: string | undefined): string {
-  const n = toFiniteNumber(value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  if (unit === "$T") return `$${n.toFixed(Math.abs(n) < 0.1 ? 3 : 2)}T`;
-  if (unit === "$bn") return `$${n.toFixed(1)}bn`;
-  return fmtValue(value, unit, unit === "bps" ? 1 : 2);
-}
-
-function fmtContracts(value: unknown): string {
-  const n = toFiniteNumber(value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  return n.toLocaleString("en-US", { maximumFractionDigits: 0 });
-}
-
-function fmtPct(value: unknown): string {
-  const n = toFiniteNumber(value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  const sign = n > 0 ? "+" : "";
-  return `${sign}${n.toFixed(1)}% OI`;
-}
-
-function fmtAuctionSize(value: unknown): string {
-  const n = toFiniteNumber(value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  return `$${n.toFixed(1)}bn`;
-}
-
-function fmtSupplyMetric(tile: SummaryTile): string {
-  const n = toFiniteNumber(tile.value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  if (tile.unit === "$T") return `$${n.toFixed(2)}T`;
-  if (tile.unit === "$bn") return `$${n.toFixed(1)}bn`;
-  if (tile.unit === "%") return `${n.toFixed(1)}%`;
-  if (tile.unit === "x") return `${n.toFixed(2)}x`;
-  return fmtValue(tile.value, tile.unit, 2);
-}
-
-function fmtPercent(value: unknown): string {
-  const n = toFiniteNumber(value, Number.NaN);
-  if (!Number.isFinite(n)) return "n/a";
-  return `${n.toFixed(1)}%`;
-}
-
-function PolicySection({ policy }: { policy: Policy }) {
-  const path = policy.implied_path ?? [];
-  return (
-    <div className={styles.policyGrid}>
-      <article className={styles.policyCard}>
-        <div className={styles.policyCardTop}>
-          <h3>Policy Rate</h3>
-          <span>FRED + Fed</span>
-        </div>
-        <dl className={styles.policyRows}>
-          {latestPolicyRows(policy).map(([label, value]) => (
-            <div key={label}>
-              <dt>{label}</dt>
-              <dd>{value}</dd>
-            </div>
-          ))}
-        </dl>
-        <p>{policy.policy_read ?? "Official policy metadata unavailable."}</p>
-      </article>
-
-      <article className={styles.policyCard}>
-        <div className={styles.policyCardTop}>
-          <h3>Market-Implied Path</h3>
-          <span>Fed funds futures</span>
-        </div>
-        {path.length ? (
-          <div className={styles.policyPathGrid}>
-            {path.slice(0, 5).map((point: PolicyPathPoint) => (
-              <div className={styles.pathPill} key={point.meeting_date}>
-                <span>{point.label}</span>
-                <strong className={policyToneClass(point.stance)}>
-                  {fmtValue(point.probability, "%", 0)}
-                </strong>
-                <small>{point.stance.toLowerCase()}</small>
-                <i>
-                  <b
-                    style={{
-                      width: `${Math.max(0, Math.min(100, toFiniteNumber(point.probability, 0)))}%`,
-                    }}
-                  />
-                </i>
-              </div>
-            ))}
-          </div>
-        ) : (
-          <div className={styles.policyMissing}>Futures path unavailable</div>
-        )}
-        <p>{policy.path_read ?? "No implied-path source is persisted yet."}</p>
-      </article>
-
-      <article className={styles.policyCard}>
-        <div className={styles.policyCardTop}>
-          <h3>Plumbing</h3>
-          <span>FRED</span>
-        </div>
-        <dl className={styles.policyRows}>
-          {(policy.plumbing ?? []).map((row) => (
-            <div key={row.label}>
-              <dt>{row.label}</dt>
-              <dd>
-                {fmtPolicyMetric(row.value, row.unit)}
-                {row.qualifier ? <small>{row.qualifier}</small> : null}
-              </dd>
-            </div>
-          ))}
-        </dl>
-        <p>{policy.plumbing_read ?? "Fed plumbing series unavailable."}</p>
-      </article>
-    </div>
-  );
-}
-
-function PositioningSection({ positioning }: { positioning: Positioning }) {
-  const rows = positioning.rows ?? [];
-  const details = positioning.details ?? [];
-  if (!rows.length && !details.length) {
-    return (
-      <div className={styles.notePanel}>
-        <strong>{statusLabel(positioning.status)}</strong>
-        <p>CFTC/TIC feeds not wired in Phase 1.</p>
-      </div>
-    );
-  }
-  return (
-    <div className={styles.positioningStack}>
-      <div className={styles.compactGrid}>
-        {rows.map((row) => (
-          <article className={styles.kpiTile} key={row.label}>
-            <span>{row.label}</span>
-            <strong>{fmtContracts(row.value)}</strong>
-            <small>{row.unit}</small>
-          </article>
-        ))}
-      </div>
-      <div className={styles.positioningTableWrap}>
-        <table className={styles.positioningTable}>
-          <thead>
-            <tr>
-              <th>Contract</th>
-              <th>Open interest</th>
-              <th>Leveraged funds</th>
-              <th>Asset managers</th>
-              <th>Dealers</th>
-            </tr>
-          </thead>
-          <tbody>
-            {details.map((row) => (
-              <tr key={`${row.contract_code}-${row.obs_date ?? ""}`}>
-                <td>
-                  <strong>{row.tenor_bucket}</strong>
-                  <small>{row.contract_name}</small>
-                </td>
-                <td>{fmtContracts(row.open_interest)}</td>
-                <td>
-                  {fmtContracts(row.lev_money_net)}
-                  <small>{fmtPct(row.lev_money_net_pct_oi)}</small>
-                </td>
-                <td>
-                  {fmtContracts(row.asset_mgr_net)}
-                  <small>{fmtPct(row.asset_mgr_net_pct_oi)}</small>
-                </td>
-                <td>
-                  {fmtContracts(row.dealer_net)}
-                  <small>{fmtPct(row.dealer_net_pct_oi)}</small>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-      <p className={styles.positioningRead}>
-        {positioning.positioning_read ??
-          "CFTC TFF Treasury futures positioning is unavailable."}
-      </p>
-    </div>
-  );
-}
-
-function SupplySection({ supply }: { supply: Supply | undefined }) {
-  const recentAuctions = supply?.recent_auctions ?? [];
-  const fiscal = supply?.fiscal ?? [];
-  if (!recentAuctions.length && !fiscal.length) {
-    return (
-      <div className={styles.notePanel}>
-        <strong>{statusLabel(supply?.status)}</strong>
-        {(supply?.notes?.length
-          ? supply.notes
-          : ["Treasury auction and FiscalData supply feeds are unavailable."]
-        ).map((note) => (
-          <p key={note}>{note}</p>
-        ))}
-      </div>
-    );
-  }
-
-  return (
-    <div className={styles.supplyGrid}>
-      <article className={styles.supplyCard}>
-        <div className={styles.policyCardTop}>
-          <h3>Recent auctions</h3>
-          <span>TreasuryDirect</span>
-        </div>
-        <div className={styles.supplyTableWrap}>
-          <table className={styles.supplyTable}>
-            <thead>
-              <tr>
-                <th>Issue</th>
-                <th>Size</th>
-                <th>High rate</th>
-                <th>Bid cover</th>
-                <th>Indirect</th>
-              </tr>
-            </thead>
-            <tbody>
-              {recentAuctions.map((row) => (
-                <tr key={`${row.cusip}-${row.auction_date}`}>
-                  <td>
-                    <strong>
-                      {row.security_term} {row.security_type}
-                    </strong>
-                    <small>{row.auction_date}</small>
-                  </td>
-                  <td>{fmtAuctionSize(row.offering_amount)}</td>
-                  <td>{fmtValue(row.high_rate, "%", 3)}</td>
-                  <td>{fmtValue(row.bid_to_cover, "", 2)}</td>
-                  <td>{fmtPercent(row.indirect_bidder_pct)}</td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-        {supply?.supply_read ? (
-          <p className={styles.positioningRead}>{supply.supply_read}</p>
-        ) : null}
-      </article>
-
-      <article className={styles.supplyCard}>
-        <div className={styles.policyCardTop}>
-          <h3>Issuance & fiscal</h3>
-          <span>FiscalData + FRED</span>
-        </div>
-        <div className={styles.supplyMetricGrid}>
-          {fiscal.map((tile) => (
-            <article className={styles.kpiTile} key={tile.label}>
-              <span>{tile.label}</span>
-              <strong>{fmtSupplyMetric(tile)}</strong>
-              <small>{statusLabel(tile.status)}</small>
-            </article>
-          ))}
-          {(supply?.auctions ?? []).map((tile) => (
-            <article className={styles.kpiTile} key={tile.label}>
-              <span>{tile.label}</span>
-              <strong>{fmtSupplyMetric(tile)}</strong>
-              <small>{statusLabel(tile.status)}</small>
-            </article>
-          ))}
-        </div>
-      </article>
-    </div>
   );
 }
